+++ server/libs/modules/components/script/src/main/java/com/bytechef/component/script/ScriptComponentDefinitionFactory.java
@@ -17,7 +17,7 @@
 package com.bytechef.component.script;
 
 import static com.bytechef.component.definition.ComponentDSL.component;
-import static com.bytechef.platform.component.constant.ScriptConstants.SCRIPT;
+import static com.bytechef.platform.component.definition.ScriptComponentDefinition.SCRIPT;
 
 import com.bytechef.component.ComponentDefinitionFactory;
 import com.bytechef.component.definition.ComponentDefinition;
