+++ server/libs/modules/components/script/src/main/java/com/bytechef/component/script/action/ScriptJavaAction.java
@@ -29,7 +29,7 @@ import static com.bytechef.component.definition.ComponentDSL.string;
 import static com.bytechef.component.definition.ComponentDSL.time;
 import static com.bytechef.component.script.constant.ScriptConstants.INPUT;
 import static com.bytechef.component.script.constant.ScriptConstants.JAVA;
-import static com.bytechef.platform.component.constant.ScriptConstants.SCRIPT;
+import static com.bytechef.platform.component.definition.ScriptComponentDefinition.SCRIPT;
 
 import com.bytechef.atlas.execution.domain.TaskExecution;
 import com.bytechef.atlas.worker.task.handler.TaskHandler;
