+++ server/libs/platform/platform-component/platform-component-api/src/main/java/com/bytechef/platform/component/definition/ScriptComponentDefinition.java
@@ -25,6 +25,11 @@ public interface ScriptComponentDefinition extends ComponentDefinition {
 
     /**
      *
+     */
+    String SCRIPT = "script";
+
+    /**
+     *
      * @return
      */
     FilterConnectionDefinitionPredicate getFilterConnectionDefinition();
